Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -4,10 +4,8 @@ int main() {
     int num1, num2;
     char operator;
 
-    // Input two numbers and the operator
     scanf("%d %c %d", &num1, &operator, &num2);
 
-    // Perform the calculation based on the operator
     switch(operator) {
         case '+':
             printf("%d\n", num1 + num2);
@@ -20,7 +18,7 @@ int main() {
             break;
         case '/':
             if (num2 != 0)
-                printf("%f\n", (float)num1 / num2);
+                printf("%d\n", num1 / num2);
             else
                 printf("Error\n");
             break;
